Added maximum scaling in fullscreen mode
When in fullscreen mode, the image is automatically scaled to the maximum
width or height possible, depending on (1) the monitor/screen ratio and (2)
the image width/height ratio: Whether we should zoom to width or height
depends on the monitor resolution ratio.
Scaling is only done, if user did not manually zoom the image in or out.